release v0.24.2

+++ src/util/BitReverse.cxx
@@ -1,21 +1,5 @@
-/*
- * Copyright 2003-2021 The Music Player Daemon Project
- * http://www.musicpd.org
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, write to the Free Software Foundation, Inc.,
- * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
- */
+// SPDX-License-Identifier: GPL-2.0-or-later
+// Copyright The Music Player Daemon Project
 
 #include "BitReverse.hxx"
 
@@ -24,7 +8,7 @@ GenerateBitReverseTable() noexcept
 {
 	BitReverseTable table{};
 	for (unsigned i = 0; i < 256; ++i)
-		table.data[i] = BitReverseMultiplyModulus(i);
+		table.data[i] = BitReverseMultiplyModulus(static_cast<std::byte>(i));
 	return table;
 }
